python/requirements.txt
python-dotenv
pandas
requests
numpy
Flask